fix: refactor LoacationHeading and SearchBar components
