+++ src/components/LocationHeading/LocationHeading.tsx
@@ -1,20 +1,13 @@
-import React, { Component } from 'react';
+import React from 'react';
+import { paths } from '../../constants/constants';
 import { withRouter, WithRouterProps } from '../../utils/withRouter';
 
-const paths: { [key: string]: string } = {
-  ['/']: 'Home',
-  ['/form']: 'Form',
-  ['/aboutus']: 'About Us',
-};
-
-class LocationHeading extends Component<WithRouterProps & { className: string }> {
-  render() {
-    return (
-      <>
-        <h2 className={this.props.className}>{paths[this.props.location.pathname]}</h2>
-      </>
-    );
-  }
+function LocationHeading(props: WithRouterProps & { className: string }) {
+  return (
+    <>
+      <h2 className={props.className}>{paths[props.location.pathname]}</h2>
+    </>
+  );
 }
 
 export default withRouter(LocationHeading);
