+++ src/constants/constants.tsx
@@ -1,0 +1,5 @@
+export const paths: { [key: string]: string } = {
+  ['/']: 'Home',
+  ['/form']: 'Form',
+  ['/aboutus']: 'About Us',
+};
